some changes

+++ src/simulator/memory/CacheLine.java
@@ -3,7 +3,7 @@ package simulator.memory;
 public class CacheLine {
     public boolean valid;
     public int tag;      // memory address (0–2047)
-    public short data;   // 16-bit word
+    public int data;   // 16-bit word
     public int age;      // FIFO count
 
     public CacheLine() {
